Replace TrapezoidProfile with TrapezoidalMotionProfile
This new class provides a generic mechanism for filtering a reference
based on a mechanism's trapezoidal velocity constraints. By profiling in
this way, refernces can be dynamically limited to what a subsystem is
capable of tracking.

This approach also generalizes to two dimensions where the goal pose of
a drivetrain can be filtered based on the drivetrain's constraints to
provide new references on the fly from the current pose.

A units library was also added to make the implementation easier to
follow. This can be expanded to other aspects of the robot's controllers
later.

Change-Id: Ie058d15409055ab892bf65c4f02d641cebd03420